Add database
Use sql to complete pagination and filtering - this should be much more
scalable than doing it in js. It does introduce the overhead of an
additional db query (to get the count of advocates given a search term).
This should be minimal overhead, and can be performed in parallel to the
actual data query.